Fix display for inputs in checkbox/radio text
A checkbox has a title and a description as well as possible extra text
to add to the title. The title and description are formatted as labels
whereas the extra text is *not* included in the label, but should be
formatted immediately after the title and before the description label.