Fix warnings on view_entry.php
The page emits warnings when the journal entry doesn't match the picture
passed in by the user. The picture has a certain date, but the journal
restricts the "list" query to the dates of the journal. The list is then
empty, but the mobile mode was still trying to show the position of the
picture within the list of other pictures (which can't work because that
picture is not in the journal's list of pictures). The fix simply avoids
emitting the "counts" when in multiple mode, conditioning on the same
variable used in non-mobile mode.